pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langsmith>=0.7.30",
    "openai>=2.31.0",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "sentence-transformers>=5.4.0",
]
